Created possible README
Most projects have a README explaining what the project is and how it can be used. I felt it would probably be helpful to have one for this code, but I'm not sure if this is exactly the message we want to send, so I'm contributing this to a separate branch and submitting a pull request to discuss the wording and contents.